vendor: fonts: add HarmonyOS Sans
HarmonyOS Sans is a font designed on the basis of end-user feedback
which includes a variety of considerations, including the dimensions
of different devices, usage scenarios, and different users'
requirements for font size and weight due to differences in
line-of-sight and angle-of-view.

Font obtained from Huawei Design Docs [1]

Licensed under HarmonyOS Sans Fonts License Agreement.

[1] https://developer.harmonyos.com/en/docs/design/des-guides/font-0000001157868583

Change-Id: Ic3a211b19735b134028bd2d9745dd015698d435c
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Reviewed-on: https://review.statixos.com/c/android_vendor_statix/+/9064
Tested-by: Anay Wadhera <anay1018@gmail.com>
Reviewed-by: Anay Wadhera <anay1018@gmail.com>